+++ app/src/main/java/org/hyperledger/besu/controller/BesuControllerBuilder.java
@@ -852,6 +852,7 @@ public abstract class BesuControllerBuilder implements MiningConfigurationOverri
             worldStateStorageCoordinator,
             ethPeers,
             snapMessages,
+            scheduler,
             synchronizer);
 
     final MiningCoordinator miningCoordinator =
@@ -1268,6 +1269,7 @@ public abstract class BesuControllerBuilder implements MiningConfigurationOverri
       final WorldStateStorageCoordinator worldStateStorageCoordinator,
       final EthPeers ethPeers,
       final EthMessages snapMessages,
+      final EthScheduler ethScheduler,
       final Synchronizer synchronizer) {
     return Optional.of(
         new SnapProtocolManager(
@@ -1275,6 +1277,7 @@ public abstract class BesuControllerBuilder implements MiningConfigurationOverri
             syncConfig.getSnapSyncConfiguration(),
             ethPeers,
             snapMessages,
+            ethScheduler,
             protocolSchedule,
             protocolContext,
             synchronizer));